medidor_lentes/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cloudinary==1.32.0
Django==4.1.7
django-cors-headers==3.3.0
django-cpf-cnpj==1.0.0
django-crispy-forms==2.0
django-discover-runner==1.0
django-filter==22.1
django-role-permissions==3.1.1
djangorestframework==3.14.0
gunicorn==20.1.0
importlib-metadata==5.0.0
Markdown==3.4.1
nodeenv==1.7.0
npm==0.1.1
numpy==1.24.2
opencv-python==4.7.0.72
opencv-python-headless==4.7.0.72
optional-django==0.1.0
Pillow==9.4.0
psycopg2==2.9.3
pytz==2020.1
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.15
whitenoise==6.4.0
xlrd==2.0.1
zipp==3.8.1
